move sentry to utils folder
