+++ src/entrypoints/api.ts
@@ -2,7 +2,7 @@ import { HttpAdapterHost, NestFactory } from '@nestjs/core'
 import { SwaggerModule } from '@nestjs/swagger'
 import { ValidationPipe, VersioningType } from '@nestjs/common'
 import { AppModule } from '../app.module.js'
-import { initSentry } from '../helpers/sentry.js'
+import { initSentry } from '../utils/sentry.js'
 import { HttpExceptionFilter } from '../utils/exceptions/http-exception.filter.js'
 import { buildDocumentationConfig } from '../helpers/documentation.js'
