+++ src/entrypoints/worker.ts
@@ -3,7 +3,7 @@ import yargs from 'yargs'
 import { hideBin } from 'yargs/helpers'
 import { AppModule } from '../app.module.js'
 import { QueueName } from '../modules/pgboss/types/queue-name.enum.js'
-import { initSentry } from '../helpers/sentry.js'
+import { initSentry } from '../utils/sentry.js'
 import { PgBossWorkerModule } from '../modules/pgboss-worker/pgboss-worker.module.js'
 
 async function bootstrap (): Promise<void> {
